// src/api/enum/api-periods.enum.ts
export enum PeriodsEnum {
  SEC_1 = '1SEC',
  SEC_2 = '2SEC',
  SEC_3 = '3SEC',
  SEC_4 = '4SEC',
  SEC_5 = '5SEC',
  SEC_6 = '6SEC',
  SEC_10 = '10SEC',
  SEC_15 = '15SEC',
  SEC_20 = '20SEC',
  SEC_30 = '30SEC',

  MIN_1 = '1MIN',
  MIN_2 = '2MIN',
  MIN_3 = '3MIN',
  MIN_4 = '4MIN',
  MIN_5 = '5MIN',
  MIN_6 = '6MIN',
  MIN_10 = '10MIN',
  MIN_15 = '15MIN',
  MIN_20 = '20MIN',
}